Move elements to separate module
